+++ httpfs/client/__main__.py
@@ -3,8 +3,8 @@ import logging
 import os
 import sys
 import yaml
-from httpfs.common.TextCredStore import Cred, CredStore
-from httpfs.common.CredStorage import TextCredStore
+from httpfs.common.CredModels import Cred, CredStore
+from httpfs.common.TextCredStore import TextCredStore
 
 from fuse import FUSE
 from httpfs.client import HttpFsClient
