+++ httpfs/common/__init__.py
@@ -1,3 +1,3 @@
 from .HttpFsRequest import HttpFsRequest
 from .HttpFsResponse import HttpFsResponse
-from .CredStorage import CredStore
+from .TextCredStore import CredStore
